move get team by id into team gateway
