+++ SocialCareCaseViewerApi/V1/Gateways/DatabaseGateway.cs
@@ -556,16 +556,6 @@ namespace SocialCareCaseViewerApi.V1.Gateways
             return teamsWorkerBelongsIn;
         }
 
-        public Team GetTeamByTeamId(int teamId)
-        {
-            return _databaseContext.Teams
-                .Where(x => x.Id == teamId)
-                .Include(x => x.WorkerTeams)
-                .ThenInclude(x => x.Worker)
-                .ThenInclude(x => x.Allocations)
-                .FirstOrDefault();
-        }
-
         public Team GetTeamByTeamName(string teamName)
         {
             return _databaseContext.Teams
